requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pymysql
cryptography
pydantic[email]
python-dotenv
